Provide a baseline mojo that can work also with maven 3.0.X
- The baseline mojo uses the new format of the aether library, the one
hosted on eclipse foundation, but that is usable only starting from
maven 3.1.X, now there are quite few projects that would still benefit
of this mojo but are still at 3.0.X level, so for those project this
fix clones the BaseLineMojo in BaseLine30Mojo that is triggered by the
goal "baseline30" so you can still get the benefit of this mojo for
3.0.X projects by simply using the goal baseline30 instead of
baseline. Definitively this one could be done more elegantly but not
being a maven mojo expert this at least gets the ball rolling.

Signed-off-by: Giovanni Meo <gmeo@cisco.com>